[CHANGE] Flight Controller. [ADD] Pegasus engine.

+++ main/PegasusProcessor.cpp
@@ -28,19 +28,19 @@ void PegasusProcessor::fillCommands(String buffer) {
       // if finished a number
       if (c == ';') {
          if (count == 0) {
-         throttle = number_str.toFloat();
+         throttle = number_str.toInt();
          number_str = "";
          count++;
       } else if (count == 1) {
-         pitch = number_str.toFloat();
+         pitch = number_str.toInt();
          number_str = "";
          count++;
       } else if (count == 2) {
-         roll = number_str.toFloat();
+         roll = number_str.toInt();
          number_str = "";
          count++;
       } else if (count == 3) {
-         yaw = number_str.toFloat();
+         yaw = number_str.toInt();
          number_str = "";
          count = 0;
       }
